requirements.txt
# Minimal runtime requirements for this assignment
transformers>=4.41.0
torch>=2.2.0; platform_system != "Darwin" or platform_machine != "arm64"
# On Apple Silicon you can use:
# pip install 'torch>=2.2.0' --index-url https://download.pytorch.org/whl/cpu
Pillow>=10.0.0
